[Fixed] In the edit the image remains saved even if in the input the file path is empty (FOR DISHES) Re-added the vote field in edit and create (FOR DISHES)

+++ app/Http/Controllers/Admin/DishController.php
@@ -98,23 +98,31 @@ class DishController extends Controller
             $form_data['slug'] = $dish->slug;
         }
 
-        //* edit per l'IMMAGINE (FUNZIONANTE ANCHE SE L'IMMAGINE è STATA PRESA DA UN API)
-        //* se è stata caricata un immagine (dal campo di input nel form) e se il value dell'input(/il percorso dell'immagine) non contiene http:// o https:// quindi l'immagine NON è STATA PRESA DALL'API
-        if (array_key_exists('image_path', $form_data) && !(str_contains($form_data['image_path'], 'http://') || str_contains($form_data['image_path'], 'https://'))){
+        $oldImagePath = $dish->image_path;
 
-            //* se l'immagine esiste (NEL DB) vuol dire che ne ho caricata una nuova E NON è STATA PRESA DALL'API (NEL DB) e quindi ELIMINO quella precedente
-            if ($dish->image_path && !(str_contains($dish->image_path, 'http://') || str_contains($dish->image_path, 'https://'))) {
-                // se è presente sul disco in public ed elimina l'immagine già presente
-                Storage::disk('public')->delete($dish->image_path);
+        // Verifica se l'input di tipo "file" per l'immagine è vuoto
+        if ($request->hasFile('image_path')) {
+                //* edit per l'IMMAGINE (FUNZIONANTE ANCHE SE L'IMMAGINE è STATA PRESA DA UN API)
+                //* se è stata caricata un immagine (dal campo di input nel form) e se il value dell'input(/il percorso dell'immagine) non contiene http:// o https:// quindi l'immagine NON è STATA PRESA DALL'API
+                if (array_key_exists('image_path', $form_data) && !(str_contains($form_data['image_path'], 'http://') || str_contains($form_data['image_path'], 'https://'))){
+
+                    //* se l'immagine esiste (NEL DB) vuol dire che ne ho caricata una nuova E NON è STATA PRESA DALL'API (NEL DB) e quindi ELIMINO quella precedente
+                    if ($dish->image_path && !(str_contains($dish->image_path, 'http://') || str_contains($dish->image_path, 'https://'))) {
+                        // se è presente sul disco in public ed elimina l'immagine già presente
+                        Storage::disk('public')->delete($dish->image_path);
+                    }
+
+                    // l'immagine viene salvata nel db con il percorso(uploads/WQZmcTvnkWAVcb35NKprvhlCQfRE2VXRTpSytd8C.jpg)
+                    $form_data['image_path'] = Storage::put('uploads', $form_data['image_path']);
+                }
+            } else {
+                // Mantieni l'immagine precedente
+                $dish->image_path = $oldImagePath;
             }
-
-            // l'immagine viene salvata nel db con il percorso(uploads/WQZmcTvnkWAVcb35NKprvhlCQfRE2VXRTpSytd8C.jpg)
-            $form_data['image_path'] = Storage::put('uploads', $form_data['image_path']);
-        }
 
         $dish->update($form_data);
 
-        return redirect()->route('admin.dishes.show', $dish);
+        return redirect()->route('admin.dishes.show', compact('dish', 'oldImagePath'));
     }
 
     /**
